Identity verification system implemented. Users only could modify, delete and see their own information
