+++ src/repositories/educationRepository.js
@@ -42,18 +42,6 @@ class EducationRepository extends BaseRepository {
     return result;
   }
 
-  async checkIdentity(id, token) {
-    const person = await this.read(id);
-    if(person[0].person_id === token.personId) {
-      return true;
-    } else {
-      const error = new Error();
-      error.User = 'Can not verify identity, id does not exist';
-      error.status = 401;
-      throw error;
-    }
-  }
-
   async deleteEducationWithForeigns(id) {
     await this.db('Utilities').where({ education_id: id }).del();
     await this.db('Aditionals').where({ education_id: id }).del();
